Switch from long to int in the Pyrex and Cython extensions to match the C and C++ ones.

+++ processSlice_cython_py3.pyx
@@ -2,10 +2,10 @@
 
 #  A Cython extensions to calculate a slice of the overall calculation of Pi using quadrature.
 #
-#  Copyright © 2010 Russel Winder
+#  Copyright © 2010--2011 Russel Winder
 
-def processSlice ( long id , long sliceSize , double delta ) :
-  cdef long i = 0
+def processSlice ( int id , int sliceSize , double delta ) :
+  cdef int i = 0
   cdef double sum = 0.0
   cdef double x
   for i in range ( 1 + id * sliceSize , ( id + 1 ) * sliceSize ) :
